Change coordinate translation apis to take doubles
Change gtk_widget_translate_coordinates and
gtk_native_get_surface_transform to operate
on doubles. Update all callers.

+++ gtk/gtkemojichooser.c
@@ -253,7 +253,7 @@ scroll_to_child (GtkWidget *child)
   GtkEmojiChooser *chooser;
   GtkAdjustment *adj;
   GtkAllocation alloc;
-  int pos;
+  double pos;
   double value;
   double page_size;
